data type alignment

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: stena-he <stena-he@student.42wolfsburg.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/27 19:12:49 by stena-he          #+#    #+#             */
-/*   Updated: 2022/07/03 19:21:39 by stena-he         ###   ########.fr       */
+/*   Updated: 2022/07/03 19:54:19 by stena-he         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,7 +37,7 @@ int		ft_put_char(va_list args);
 int		ft_put_str(va_list args);
 // pointers -p
 int		ft_put_ptr(va_list args);
-void	ft_put_pointer(unsigned long long ptr);
+void	ft_put_pointer(unsigned long ptr);
 // integers & decimals -i -d
 int		ft_put_nbr(va_list args);
 void	ft_putnbr(int nb);
@@ -48,9 +48,9 @@ void	ft_put_unsign_nb(unsigned int nb);
 int		ft_unsign_length(unsigned int x);
 // hexadecimals -x -X
 int		ft_put_low_hexadec(va_list args);
-void	ft_put_low_hexadec_nb(unsigned long long int nb);
+void	ft_put_low_hexadec_nb(unsigned long int nb);
 int		ft_put_upp_hexadec(va_list args);
 void	ft_put_upp_hexadec_nb(unsigned int nb);
-int		ft_hexadec_length(unsigned long long x);
+int		ft_hexadec_length(unsigned long x);
 
 #endif
